strip homebrew-fork utils.rb
* remove unused methods gzip and paths
* migrate ohai, opoo, onoe into our own codebase
* move GitHub module into a separate file
* remove unused requires
